Phpstan fixes (#206)
* remove positive_int annotations

will fail on a 0 value

* alpha values can be floats

* increase phpstan level + fixes

* Fix styling

* exclude relevant phpstan issues from baseline

---------

+++ src/Drivers/Color.php
@@ -10,7 +10,7 @@ abstract class Color
 {
     abstract public function initFromInteger(int $value): self;
 
-    /** @param  array<positive-int>  $value */
+    /** @param  array<int>  $value */
     abstract public function initFromArray(array $value): self;
 
     abstract public function initFromString(string $value): self;
@@ -25,7 +25,7 @@ abstract class Color
 
     abstract public function getHex(string $prefix): string;
 
-    /** @return array<positive-int> */
+    /** @return array<int, float|int> */
     abstract public function getArray(): array;
 
     abstract public function getRgba(): string;
@@ -91,7 +91,7 @@ abstract class Color
             $result[0] = ($matches[1] >= 0 && $matches[1] <= 255) ? (int) ($matches[1]) : 0;
             $result[1] = ($matches[2] >= 0 && $matches[2] <= 255) ? (int) ($matches[2]) : 0;
             $result[2] = ($matches[3] >= 0 && $matches[3] <= 255) ? (int) ($matches[3]) : 0;
-            $result[3] = ($matches[4] >= 0 && $matches[4] <= 1) ? $matches[4] : 0;
+            $result[3] = ($matches[4] >= 0 && $matches[4] <= 1) ? (int) ($matches[4]) : 0;
         } else {
             throw InvalidColor::make($colorValue);
         }
